requirements.txt
supabase
openai
yfinance
newsapi-python
feedparser
google-generativeai
